# apps/web/docs/openapi/openapi.merged.yaml
openapi: 3.0.3
info:
  title: Africonnect B2B — REST + Convex HTTP API
  version: 0.1.0
  description: ""
servers:
  - url: http://localhost:3000
    description: Next.js app — use for paths tagged **Next.js REST** (`/api/...` except Convex-only hosts).
  - url: https://hallowed-squirrel-738.convex.cloud
...
paths:
...
  /api/payments/webhook:
    get:
      tags:
        - Next.js REST
      summary: GET /api/payments/webhook
      description: |-
        Handler: `app/api/payments/webhook/route.ts`.

        Call this URL **on the Next.js server** (see `servers` entry for the app), not on the Convex deployment.
      responses:
        "200":
          description: Success — response shape is defined by the route implementation.
        "400":
          description: Bad request (route-specific validation).
        "401":
          description: Unauthorized (when Clerk/session applies).
        "500":
          description: Server error.
    post:
      tags:
        - Next.js REST
      summary: POST /api/payments/webhook
      description: |-
        Handler: `app/api/payments/webhook/route.ts`.

        Call this URL **on the Next.js server** (see `servers` entry for the app), not on the Convex deployment.
      responses:
        "200":
          description: Success — response shape is defined by the route implementation.
        "400":
          description: Bad request (route-specific validation).
        "401":
          description: Unauthorized (when Clerk/session applies).
        "500":
          description: Server error.
...
  /api/payouts/webhook:
    post:
      tags:
        - Next.js REST
      summary: POST /api/payouts/webhook
      description: |-
        Handler: `app/api/payouts/webhook/route.ts`.

        Call this URL **on the Next.js server** (see `servers` entry for the app), not on the Convex deployment.
      responses:
        "200":
          description: Success — response shape is defined by the route implementation.
        "400":
          description: Bad request (route-specific validation).
        "401":
          description: Unauthorized (when Clerk/session applies).
        "500":
          description: Server error.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token of the format "Bearer {token}" for normal authentication and "Convex {token}" for admin tokens.
...